requirements.txt
cryptography
pypdf
pytest
pillow
